requirements.txt
flask
pyrebase4
gunicorn
requests==2.32.3
urllib3==1.26.20
setuptools